refactor(drawer): fix drawer
